Annotate legacy comments files and v1 api models with 'deprecate' prefix
