+++ app/src/main/AndroidManifest.xml
@@ -50,7 +50,7 @@
             android:name=".ui.activities.EditProfileActivity"
             android:theme="@style/SettingsActivity" />
         <activity
-            android:name=".ui.activities.CommentsActivity"
+            android:name=".ui.activities.DeprecatedCommentsActivity"
             android:parentActivityName=".ui.activities.ProjectActivity"
             android:theme="@style/CommentsActivity" />
         <activity android:name=".ui.activities.CreatorBioActivity" />
