+++ app/src/main/java/com/kickstarter/libs/utils/CommentUtils.java
@@ -1,6 +1,6 @@
 package com.kickstarter.libs.utils;
 
-import com.kickstarter.models.Comment;
+import com.kickstarter.models.DeprecatedComment;
 import com.kickstarter.models.User;
 
 import androidx.annotation.NonNull;
@@ -11,11 +11,11 @@ public final class CommentUtils {
     throw new AssertionError();
   }
 
-  public static boolean isUserAuthor(final @NonNull Comment comment, final @Nullable User user) {
+  public static boolean isUserAuthor(final @NonNull DeprecatedComment comment, final @Nullable User user) {
     return user != null && comment.author().id() == user.id();
   }
 
-  public static boolean isDeleted(final @NonNull Comment comment) {
+  public static boolean isDeleted(final @NonNull DeprecatedComment comment) {
     return !DateTimeUtils.isEpoch(comment.deletedAt());
   }
 }
