+++ app/src/main/java/com/kickstarter/services/ApiClientType.java
@@ -4,7 +4,7 @@ import com.google.gson.JsonObject;
 import com.kickstarter.libs.Config;
 import com.kickstarter.models.Backing;
 import com.kickstarter.models.Category;
-import com.kickstarter.models.Comment;
+import com.kickstarter.models.DeprecatedComment;
 import com.kickstarter.models.Location;
 import com.kickstarter.models.Message;
 import com.kickstarter.models.MessageThread;
@@ -16,7 +16,7 @@ import com.kickstarter.models.Update;
 import com.kickstarter.models.User;
 import com.kickstarter.services.apiresponses.AccessTokenEnvelope;
 import com.kickstarter.services.apiresponses.ActivityEnvelope;
-import com.kickstarter.services.apiresponses.CommentsEnvelope;
+import com.kickstarter.services.apiresponses.DeprecatedCommentsEnvelope;
 import com.kickstarter.services.apiresponses.DiscoverEnvelope;
 import com.kickstarter.services.apiresponses.EmailVerificationEnvelope;
 import com.kickstarter.services.apiresponses.MessageThreadEnvelope;
@@ -69,11 +69,11 @@ public interface ApiClientType {
 
   @NonNull Observable<Backing> fetchProjectBacking(final @NonNull Project project, final @NonNull User user);
 
-  @NonNull Observable<CommentsEnvelope> fetchComments(final @NonNull Project project);
+  @NonNull Observable<DeprecatedCommentsEnvelope> fetchComments(final @NonNull Project project);
 
-  @NonNull Observable<CommentsEnvelope> fetchComments(final @NonNull String paginationPath);
+  @NonNull Observable<DeprecatedCommentsEnvelope> fetchComments(final @NonNull String paginationPath);
 
-  @NonNull Observable<CommentsEnvelope> fetchComments(final @NonNull Update update);
+  @NonNull Observable<DeprecatedCommentsEnvelope> fetchComments(final @NonNull Update update);
 
   @NonNull Observable<MessageThreadEnvelope> fetchMessagesForBacking(final @NonNull Backing backing);
 
@@ -107,9 +107,9 @@ public interface ApiClientType {
 
   @NonNull Observable<Backing> postBacking(final @NonNull Project project, final @NonNull Backing backing, final boolean checked);
 
-  @NonNull Observable<Comment> postComment(final @NonNull Project project, final @NonNull String body);
+  @NonNull Observable<DeprecatedComment> postComment(final @NonNull Project project, final @NonNull String body);
 
-  @NonNull Observable<Comment> postComment(final @NonNull Update update, final @NonNull String body);
+  @NonNull Observable<DeprecatedComment> postComment(final @NonNull Update update, final @NonNull String body);
 
   @NonNull Observable<JsonObject> registerPushToken(final @NonNull String token);
