+++ app/src/main/java/com/kickstarter/services/apirequests/DeprecatedCommentBody.java
@@ -8,17 +8,17 @@ import auto.parcel.AutoParcel;
 
 @AutoGson
 @AutoParcel
-public abstract class CommentBody implements Parcelable {
+public abstract class DeprecatedCommentBody implements Parcelable {
   public abstract String body();
 
   @AutoParcel.Builder
   public abstract static class Builder {
     public abstract Builder body(String __);
-    public abstract CommentBody build();
+    public abstract DeprecatedCommentBody build();
   }
 
   public static Builder builder() {
-    return new AutoParcel_CommentBody.Builder();
+    return new AutoParcel_DeprecatedCommentBody.Builder();
   }
 
   public abstract Builder toBuilder();
