+++ app/src/main/java/com/kickstarter/ui/adapters/DeprecatedCommentsAdapter.java
@@ -11,11 +11,11 @@ import com.kickstarter.R;
 import com.kickstarter.databinding.CommentCardViewBinding;
 import com.kickstarter.databinding.EmptyCommentsLayoutBinding;
 import com.kickstarter.databinding.ProjectContextViewBinding;
-import com.kickstarter.models.Comment;
+import com.kickstarter.models.DeprecatedComment;
 import com.kickstarter.models.Project;
 import com.kickstarter.models.User;
 import com.kickstarter.ui.adapters.data.CommentsData;
-import com.kickstarter.ui.viewholders.CommentViewHolder;
+import com.kickstarter.ui.viewholders.DeprecatedCommentViewHolder;
 import com.kickstarter.ui.viewholders.EmptyCommentsViewHolder;
 import com.kickstarter.ui.viewholders.KSViewHolder;
 import com.kickstarter.ui.viewholders.ProjectContextViewHolder;
@@ -25,12 +25,12 @@ import java.util.List;
 
 import rx.Observable;
 
-public final class CommentsAdapter extends KSAdapter {
+public final class DeprecatedCommentsAdapter extends KSAdapter {
   private final Delegate delegate;
 
   public interface Delegate extends ProjectContextViewHolder.Delegate, EmptyCommentsViewHolder.Delegate {}
 
-  public CommentsAdapter(final @NonNull Delegate delegate) {
+  public DeprecatedCommentsAdapter(final @NonNull Delegate delegate) {
     this.delegate = delegate;
   }
 
@@ -47,7 +47,7 @@ public final class CommentsAdapter extends KSAdapter {
 
   public void takeData(final @NonNull CommentsData data) {
     final Project project = data.project();
-    final List<Comment> comments = data.comments();
+    final List<DeprecatedComment> comments = data.comments();
     final User user = data.user();
 
     sections().clear();
@@ -75,7 +75,7 @@ public final class CommentsAdapter extends KSAdapter {
     } else if (layout == R.layout.empty_comments_layout) {
       return new EmptyCommentsViewHolder(EmptyCommentsLayoutBinding.inflate(LayoutInflater.from(viewGroup.getContext()), viewGroup, false), this.delegate);
     } else {
-      return new CommentViewHolder(CommentCardViewBinding.inflate(LayoutInflater.from(viewGroup.getContext()), viewGroup, false));
+      return new DeprecatedCommentViewHolder(CommentCardViewBinding.inflate(LayoutInflater.from(viewGroup.getContext()), viewGroup, false));
     }
   }
