+++ app/src/main/java/com/kickstarter/ui/adapters/data/CommentsData.java
@@ -1,6 +1,6 @@
 package com.kickstarter.ui.adapters.data;
 
-import com.kickstarter.models.Comment;
+import com.kickstarter.models.DeprecatedComment;
 import com.kickstarter.models.Project;
 import com.kickstarter.models.User;
 
@@ -13,13 +13,13 @@ import auto.parcel.AutoParcel;
 @AutoParcel
 public abstract class CommentsData {
   public abstract @NonNull Project project();
-  public abstract @Nullable List<Comment> comments();
+  public abstract @Nullable List<DeprecatedComment> comments();
   public abstract @Nullable User user();
 
   @AutoParcel.Builder
   public abstract static class Builder {
     public abstract Builder project(Project __);
-    public abstract Builder comments(List<Comment> __);
+    public abstract Builder comments(List<DeprecatedComment> __);
     public abstract Builder user(User __);
     public abstract CommentsData build();
   }
@@ -31,7 +31,7 @@ public abstract class CommentsData {
   public abstract Builder toBuilder();
 
   public static @NonNull CommentsData deriveData(final @NonNull Project project,
-    final @Nullable List<Comment> comments, final @Nullable User user) {
+                                                 final @Nullable List<DeprecatedComment> comments, final @Nullable User user) {
 
     return CommentsData.builder()
       .project(project)
