+++ app/src/main/java/com/kickstarter/viewmodels/ProjectViewModel.kt
@@ -216,7 +216,7 @@ interface ProjectViewModel {
         /** Emits when we should start the campaign [com.kickstarter.ui.activities.CampaignDetailsActivity].  */
         fun startCampaignWebViewActivity(): Observable<ProjectData>
 
-        /** Emits when we should start [com.kickstarter.ui.activities.CommentsActivity].  */
+        /** Emits when we should start [com.kickstarter.ui.activities.DeprecatedCommentsActivity].  */
         fun startCommentsActivity(): Observable<Pair<Project, ProjectData>>
 
         /** Emits when we should start the creator bio [com.kickstarter.ui.activities.CreatorBioActivity].  */
